% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_ceramic_and_stone_artefact_mass}
\alias{plot_ceramic_and_stone_artefact_mass}
\title{plot_ceramic_and_stone_artefact_mass}
\usage{
plot_ceramic_and_stone_artefact_mass(the_data, calibrated_dates)
}
\description{
plot_ceramic_and_stone_artefact_mass
}

